necktie/tests/requirements.txt
-r ../requirements.txt

pytest==5.4.3
pytest-django==3.9.0
pytest-cov==2.10.0
six==1.16.0
